[jest-esbuild-transformer] Add tsnode-like runner
